use caching in get_station_data simplify suggest_names print some info-output in cli

+++ station_data.py
@@ -9,19 +9,14 @@ def unicode_csv_reader(latin1_data, **kwargs):
         yield [unicode(cell, "latin-1") for cell in row]
 
 
-def get_station_data():
-    filename = "/tmp/station_list.txt"
-    # remove old filename
-    try:
-        os.remove(filename)
-    except OSError:
-        pass
-    # write stations_list_soil.txt into filename
-    with open(filename,'wb') as file:
-        ftp = FTP("ftp-cdc.dwd.de")
-        ftp.login()
-        ftp.retrbinary('RETR /pub/CDC/help/stations_list_soil.txt', file.write)
-        ftp.quit()
+def get_station_data(filename="/tmp/station_list.txt"):
+    if not os.path.isfile(filename):
+        # write stations_list_soil.txt into filename
+        with open(filename,'wb') as file:
+            ftp = FTP("ftp-cdc.dwd.de")
+            ftp.login()
+            ftp.retrbinary('RETR /pub/CDC/help/stations_list_soil.txt', file.write)
+            ftp.quit()
     id_idx = 0
     name2id = {}
     id2meta = {}
@@ -55,7 +50,7 @@ def get_daily_recent():
     id2file = {}
     for l in ls:
         try:
-            id2file[int(l.split("_")[2])] = "ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/daily/kl/recent/" + l
+            id2file[int(l.split("_")[2])] = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/daily/kl/recent/" + l
         except:
             continue
 
@@ -82,8 +77,7 @@ def get_daily_hist():
 
 
 def suggest_names(name, name2id):
-    station_names=sorted(list(name2id.keys()))
-    return [st for st in station_names if unicode(name,"utf8").lower() in st.lower()]
+    return [st for st in name2id.keys() if unicode(name,"utf8").lower() in st.lower()]
 
 
 def get_name(name2id):
@@ -104,7 +98,9 @@ def cli():
     name2id, id2meta = get_station_data()
     id2recent = get_daily_recent()
     id2hist = get_daily_hist()
-    
+    print "# Stations: ", len(name2id)
+    print "# recent: ", len(id2recent)
+    print "# hist: ", len(id2hist)
     station_name = get_name(name2id)
     
     station_id = name2id[station_name]
